% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_sim.R
\name{baseline}
\alias{baseline}
\title{Generate baseline features for simulation}
\usage{
baseline(nsamp, ntax, corr_struct = NULL, .f)
}
\arguments{
\item{nsamp}{Integer, number of samples of the data set}

\item{ntax}{Integer, number of taxa (or features) of the data set}

\item{corr_struct}{String, correlation structure}

\item{.f}{Function, function factory to generate baseline}
}
\description{
Generate baseline features for simulation
}
