% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{data<-}
\alias{data<-}
\title{Generic to set the data slot}
\usage{
data(object) <- value
}
\arguments{
\item{object}{The simulation object}

\item{value}{value to replace}
}
\description{
Generic to set the data slot
}
