% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{null_sim}
\alias{null_sim}
\title{Function to simulate abundances}
\usage{
null_sim(n_tax, n_samp, sparsity, phi = NULL, mu = NULL)
}
\arguments{
\item{n_tax}{Number of taxa,}

\item{n_samp}{sample size}

\item{sparsity}{Sparsity}

\item{phi}{Dispersion Parameter}

\item{mu}{Mean abundance}
}
\description{
Function to get simulated counts with different sparsity levels
following a zero-inflated negative binomial distribution
}
