% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{simulation-class}
\alias{simulation-class}
\alias{data,simulation-method}
\alias{data<-,simulation-method}
\title{An S4 class to represent a simulation object}
\usage{
\S4method{data}{simulation}(x)

\S4method{data}{simulation}(object) <- value
}
\arguments{
\item{object}{Object of \code{simulation} type}

\item{value}{Value to replace, is a named list}
}
\description{
An S4 class to represent a simulation object

Setting param values
}
\section{Methods (by generic)}{
\itemize{
\item \code{data}: Accessing data slot

\item \code{data<-}: Assigning data slot
}}

\section{Slots}{

\describe{
\item{\code{data}}{A list of data.frame type objects}

\item{\code{param}}{A named list of parameters and their values}

\item{\code{type}}{A character vector of the steps of simulation}
}}

