% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{simulation}
\alias{simulation}
\title{Creating a simulation object}
\usage{
simulation(data, param, type)
}
\arguments{
\item{data}{List of data.frame type}

\item{param}{Named list of numeric or character type}

\item{type}{Vector contains all the simulation types done on the data}
}
\description{
Creating a simulation object
}
